xmatch-service/pyproject.toml
[tool.poetry]
name = "xmatch-service"
version = "24.3.0-rc1"
description = ""
authors = ["Diego Rodriguez <diegorodriguezmancini@gmail.com>"]
readme = "README.md"
packages = [
    { include = "api", from = "src" },
    { include = "core", from = "src" },
    { include = "database", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
healpy = "^1.16.6"


[tool.poetry.group.test.dependencies]
pytest-docker = "^3.1.1"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
